Implement item hiding in more idiomatic way
This commits implements todo items hiding according to filter state
in a more (i believe) idiomatic way.

Instead of relying on css class to show/hide particular DOM element,
when we specify parent component (TodoList) we first filter through
app items and call child component (Item) only when it's not hidden
according to our display logic (determined by hidden? function).

It has a nice side-effect that we don't have to pass [item filter]
tuple to Item component as just item us sufficient and also fixes
the bug that whenever our filter state was set to 'complete' and we
added new todo item, this item was shown in the list even when it
should really be hidden in my opinion.